Bug fixes
remaining:

1) * key on keypad coloured even though no key associated with it now
(easy fix)
2) Re-enable changing control scheme in demo
3) bug where shake and slowdown extend for duration of firing animation
when hit has reappeared. When the player is firing and is then hit, the
shake associated with being hit lasts much longer - for the duration of
firing.

5) energy shield and adaptive armour bars don't have "caps" - the
meniscus - when emptying
6) auto flare bar shrinks instead of cropping. The correct look would be
a combination of 5 and 6 - bar is cropped down when damage is taken like
the armour and shield bars do, but with a meniscus like on the auto
flare
7) Warning light doesn't flash when overheating in heat bar. Technically
this bug has existed ever since it was implemented, but I haven't
mentioned it before!
8) Game moves super slowly if you exit to menu and restart the level